swagger_server/controllers/receipts_controller.py
import connexion
import six

from swagger_server.models.error import Error  # noqa: E501
from swagger_server.models.receipt import Receipt  # noqa: E501
from swagger_server.models.receipt_create_request import ReceiptCreateRequest  # noqa: E501
from swagger_server.models.receipt_list import ReceiptList  # noqa: E501
from swagger_server.models.receipt_update_request import ReceiptUpdateRequest  # noqa: E501
from swagger_server import util


def receipts_delete(receipt_id):  # noqa: E501
    """Delete a receipt

    Delete a receipt by ID. # noqa: E501

    :param receipt_id: ID of the receipt to delete
    :type receipt_id: str

    :rtype: None
    """
    return 'do some magic!'


def receipts_get():  # noqa: E501
    """Retrieve all receipts

    Get a list of all receipts. # noqa: E501


    :rtype: ReceiptList
    """
    return 'do some magic!'


def receipts_post(body):  # noqa: E501
    """Create a new receipt

    Add a new receipt to the system. # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: Receipt
    """
    if connexion.request.is_json:
        body = ReceiptCreateRequest.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def receipts_receipt_id_put(body, receipt_id):  # noqa: E501
    """Update receipt details

    Update details of an existing receipt by ID. # noqa: E501

    :param body: 
    :type body: dict | bytes
    :param receipt_id: ID of the receipt to update
    :type receipt_id: str

    :rtype: Receipt
    """
    if connexion.request.is_json:
        body = ReceiptUpdateRequest.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'
